# backend/openapi/TidySense.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TidySense | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageDtoOfTaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageDtoOfTaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageDtoOfTaskDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}/complete": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTaskRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTaskRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}/drop": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DropTaskRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DropTaskRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DropTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}/carry": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarryTaskRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CarryTaskRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CarryTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}/restore": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreTaskRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreTaskRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiProblemDto": {
        "required": [
          "type",
          "title",
          "status",
          "code",
          "messageKey",
          "retryable",
          "traceId",
          "correlationId",
          "detail",
          "errors",
          "entityId",
          "expectedVersion",
          "currentVersion"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "messageKey": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "traceId": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "entityId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "currentVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "CarryTaskRequest": {
        "required": [
          "expectedVersion",
          "plannedDate"
        ],
        "type": "object",
        "properties": {
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "plannedDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "CompleteTaskRequest": {
        "required": [
          "expectedVersion",
          "completedForLocalDate"
        ],
        "type": "object",
        "properties": {
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "completedForLocalDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "CreateTaskRequest": {
        "required": [
          "title",
          "description",
          "goalId",
          "projectId",
          "plannedDate",
          "deadline",
          "sequenceId",
          "sequenceOrder"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "goalId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "projectId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "plannedDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "deadline": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "sequenceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "sequenceOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "isProtected": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "CursorPageDtoOfTaskDto": {
        "required": [
          "items",
          "page"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageInfoDto"
          }
        }
      },
...
      "DropTaskRequest": {
        "required": [
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "PageInfoDto": {
        "required": [
          "nextCursor",
          "hasMore"
        ],
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
      "TaskDto": {
        "required": [
          "id",
          "goalId",
          "projectId",
          "title",
          "description",
          "status",
          "plannedDate",
          "deadline",
          "sequenceId",
          "sequenceOrder",
          "isBlocked",
          "blockedBy",
          "isProtected",
          "carryCount",
          "completedForLocalDate",
          "source",
          "version",
          "createdAt",
          "updatedAt",
          "terminalAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "goalId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "projectId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          },
          "plannedDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "deadline": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "sequenceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "sequenceOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "isBlocked": {
            "type": "boolean"
          },
          "blockedBy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDependencyDto"
            }
          },
          "isProtected": {
            "type": "boolean"
          },
          "carryCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "completedForLocalDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "source": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "terminalAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "UpdateTaskRequest": {
        "required": [
          "title",
          "description",
          "goalId",
          "projectId",
          "plannedDate",
          "deadline",
          "sequenceId",
          "sequenceOrder",
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "goalId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "projectId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "plannedDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "deadline": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "sequenceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "sequenceOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "isProtected": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
    }
  },
...
}